+++ src/Frontend/src/components/messages2/SagaDiagram/SagaPluginNeeded.vue
@@ -1,0 +1,70 @@
+<script setup lang="ts">
+import NoSagaIcon from "@/assets/NoSaga.svg";
+import CopyClipboardIcon from "@/assets/Shell_CopyClipboard.svg";
+</script>
+
+<template>
+  <div class="body" role="status" aria-label="saga-plugin-needed">
+    <div class="saga-message">
+      <div class="saga-message-container">
+        <img class="saga-message-image" :src="NoSagaIcon" alt="" />
+        <h1 class="saga-message-title">Saga audit plugin needed to visualize saga</h1>
+        <div class="saga-message-box">
+          <p class="saga-message-text">To visualize your saga, please install the appropriate nuget package in your endpoint</p>
+          <a href="https://www.nuget.org/packages/NServiceBus.SagaAudit" class="saga-message-link">install-package NServiceBus.SagaAudit</a>
+          <img class="saga-message-icon" :src="CopyClipboardIcon" alt="" />
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
+
+<style scoped>
+.body {
+  display: flex;
+  flex: 1;
+  justify-content: center;
+}
+
+.saga-message {
+  display: flex;
+  align-items: center;
+}
+
+.saga-message-image {
+  display: block;
+  margin: auto;
+}
+
+.saga-message-title {
+  text-align: center;
+  margin: 0.2rem 0 0;
+  font-size: 1.5rem;
+  color: #cccccc;
+}
+
+.saga-message-box {
+  margin-top: 1rem;
+  padding: 1.5rem 2rem;
+  color: #535353;
+  font-size: 1rem;
+  font-weight: 900;
+  text-align: center;
+  background-color: #e4e4e4;
+}
+
+.saga-message-text {
+  margin: 0;
+}
+
+.saga-message-link {
+  font-family: "Courier New", Courier, monospace;
+  color: #aaaaaa;
+}
+
+.saga-message-icon {
+  display: inline-block;
+  margin-left: 0.5rem;
+  width: 1.5rem;
+}
+</style>
