+++ src/Frontend/src/components/messages2/SagaDiagram/SagaTimeoutMessage.vue
@@ -1,0 +1,169 @@
+<script setup lang="ts">
+import { SagaTimeoutMessage } from "./useSagaDiagramParser";
+import MessageDataBox from "./MessageDataBox.vue";
+import TimeoutIcon from "@/assets/TimeoutIcon.svg";
+import SagaTimeoutIcon from "@/assets/SagaTimeoutIcon.svg";
+
+defineProps<{
+  message: SagaTimeoutMessage;
+  isLastMessage: boolean;
+}>();
+</script>
+
+<template>
+  <div class="row row--right">
+    <div class="cell cell--center">
+      <div class="cell-inner cell-inner-line">
+        <img class="saga-icon saga-icon--center-cell saga-icon--overlap" :src="SagaTimeoutIcon" alt="" />
+        <a class="timeout-status" href="" aria-label="timeout requested">Timeout Requested = {{ message.TimeoutFriendly }}</a>
+      </div>
+    </div>
+    <div class="cell cell--side"></div>
+    <div class="cell cell--center cell--top-border">
+      <div class="cell-inner cell-inner-top"></div>
+      <div v-if="!isLastMessage" class="cell-inner cell-inner-line"></div>
+    </div>
+    <div class="cell cell--side">
+      <div class="cell-inner cell-inner-right"></div>
+      <div class="cell-inner cell-inner-side cell-inner-side--active">
+        <img class="saga-icon saga-icon--side-cell" :src="TimeoutIcon" alt="" />
+        <h2 class="message-title" aria-label="timeout message type">{{ message.MessageFriendlyTypeName }}</h2>
+        <div class="timestamp" aria-label="timeout message timestamp">{{ message.FormattedTimeSent }}</div>
+      </div>
+      <div class="message-data message-data--active">
+        <MessageDataBox />
+      </div>
+    </div>
+  </div>
+</template>
+
+<style scoped>
+.row {
+  display: flex;
+}
+
+.row--right {
+  justify-content: right;
+}
+
+.cell {
+  padding: 0;
+}
+
+.cell--center {
+  width: 50%;
+  background-color: #f2f2f2;
+  border: 0;
+}
+
+.cell--side {
+  align-self: flex-end;
+  width: 25%;
+  padding: 0;
+}
+
+.cell--top-border {
+  display: flex;
+  flex-direction: column;
+}
+
+.cell-inner-top {
+  border-top: solid 2px #000000;
+  margin-left: 1rem;
+}
+
+.cell-inner-line {
+  flex-grow: 1;
+  padding: 0.25rem 0.5rem;
+  border-left: solid 2px #000000;
+  margin-left: 1rem;
+}
+
+.cell-inner-side {
+  padding: 0.25rem 0.25rem 0;
+  border: solid 2px #cccccc;
+  background-color: #cccccc;
+}
+
+.cell-inner-side--active {
+  border: solid 2px #000000;
+}
+
+.cell-inner-right {
+  position: relative;
+  min-height: 2.5rem;
+  border: solid 2px #000000;
+  border-left: 0;
+  border-bottom: 0;
+  margin-right: 50%;
+}
+
+.cell-inner-right:after {
+  position: absolute;
+  display: block;
+  content: "";
+  border: solid 6px #000000;
+  border-top-width: 10px;
+  border-left-color: transparent;
+  border-right-color: transparent;
+  border-bottom: 0;
+  bottom: 0;
+  margin-left: 100%;
+  left: -5px;
+}
+
+.message-title {
+  margin: 0;
+  font-size: 0.9rem;
+  font-weight: 900;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
+
+.timestamp {
+  font-size: 0.9rem;
+}
+
+.message-data {
+  display: none;
+  padding: 0.2rem;
+  background-color: #ffffff;
+  border: solid 1px #cccccc;
+  font-size: 0.75rem;
+}
+
+.message-data--active {
+  display: block;
+}
+
+.timeout-status {
+  display: inline-block;
+  font-size: 1rem;
+  font-weight: 900;
+}
+
+.saga-icon {
+  display: block;
+  float: left;
+  margin-right: 0.35rem;
+}
+
+.saga-icon--side-cell {
+  width: 2rem;
+  height: 2rem;
+  padding: 0.23rem;
+}
+
+.saga-icon--center-cell {
+  float: none;
+  display: inline;
+  width: 1rem;
+  height: 1rem;
+  margin-top: -0.3rem;
+}
+
+.saga-icon--overlap {
+  margin-left: -1rem;
+}
+</style>
